+++ Models/DiemCongTacXaHoi.cs
@@ -7,7 +7,7 @@ public partial class DiemCongTacXaHoi
 {
     public string MaSinhVien { get; set; } = null!;
 
-    public int? TongDiem { get; set; }
+    public float? TongDiem { get; set; }
 
     public virtual SinhVien MaSinhVienNavigation { get; set; } = null!;
 }
